modify the style.css
remove flex on left and right links instead just flex the ul, add padding to the header and footer.
remove the margin auto on header and footer and added space-between for spacing between links.
remove the margin auto on content class, instead use
flex: 1 for the content to grow on the space available and centered
using align items and justify content.